Always set data descriptor bit flag
7zip refuses to extract zip archives if the general purpose bit flags
set in the local file header and central directory file header are not
equal. To resolve this, the data descriptor bit flag is set regardless
of whether or not `self.data_descriptor` is `None` in the `write` method
for `FileHeader`.

This does not strictly adhere to the ZIP standard, but setting the data
descriptor bit flag in the central directory file header as well as in
the local file header did not make the archive unreadable in my testing.
(I tried extracting the resulting archive with p7zip and unzip on linux)