Add Zip64 support
Change the API for adding to an archive to include an `is_zip64`
parameter. If set, the relevant fields will be stored in zip64 format.
`is_zip64` will be enabled automatically if the current size of the
archive exceeds 2^32 bytes. When writing the end of central directory
record, Zip64 will be used if any of the files were stored in Zip64
format or if the number of central directory records exceeds 2^16.

An archive using Zip64 produced using these changes was able to be
extracted by unzip, p7zip, Windows Explorer and winRAR.